# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type SloperaWaitlistInputProductId = typeof SloperaWaitlistInputProductId[keyof typeof SloperaWaitlistInputProductId];


export const SloperaWaitlistInputProductId = {
  vinyl: 'vinyl',
  cd: 'cd',
  cassette: 'cassette',
  'box-set': 'box-set',
  poster: 'poster',
  tee: 'tee',
} as const;

export interface SloperaWaitlistInput {
  /** @maxLength 254 */
  email: string;
  productId: SloperaWaitlistInputProductId;
  requestId: string;
  /** @maxLength 200 */
  website?: string;
}

export interface SloperaTragedyInput {
  /**
     * @minLength 1
     * @maxLength 5000
     */
  tragedy: string;
  requestId: string;
  /** @maxLength 100 */
  handle?: string;
  /** @maxLength 254 */
  email?: string;
  /** @maxLength 200 */
  website?: string;
}

export interface SloperaSuccess {
  message: string;
}

export interface SloperaError {
  error: string;
}