# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: slopera
    description: Slopera House email operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /slopera/waitlist:
    post:
      operationId: joinSloperaWaitlist
      tags: [slopera]
      summary: Join the Slopera House waitlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SloperaWaitlistInput"
      responses:
        "200":
          description: Waitlist signup received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SloperaSuccess"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SloperaError"
        "429":
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SloperaError"
        "503":
          description: Email delivery unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SloperaError"
  /slopera/tragedies:
    post:
      operationId: submitSloperaTragedy
      tags: [slopera]
      summary: Submit a tragedy to the Committee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SloperaTragedyInput"
      responses:
        "200":
          description: Tragedy received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SloperaSuccess"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SloperaError"
        "429":
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SloperaError"
        "503":
          description: Email delivery unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SloperaError"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    SloperaWaitlistInput:
      type: object
      required:
        - email
        - productId
        - requestId
      properties:
        email:
          type: string
          format: email
          maxLength: 254
        productId:
          type: string
          enum: [vinyl, cd, cassette, box-set, poster, tee]
        requestId:
          type: string
          format: uuid
        website:
          type: string
          maxLength: 200
    SloperaTragedyInput:
      type: object
      required:
        - tragedy
        - requestId
      properties:
        tragedy:
          type: string
          minLength: 1
          maxLength: 5000
        requestId:
          type: string
          format: uuid
        handle:
          type: string
          maxLength: 100
        email:
          type: string
          format: email
          maxLength: 254
        website:
          type: string
          maxLength: 200
    SloperaSuccess:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    SloperaError:
      type: object
      required:
        - error
      properties:
        error:
          type: string