# lib/api-zod/src/generated/types/sloperaTragedyInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SloperaTragedyInput {
  /**
     * @minLength 1
     * @maxLength 5000
     */
  tragedy: string;
  requestId: string;
  /** @maxLength 100 */
  handle?: string;
  /** @maxLength 254 */
  email?: string;
  /** @maxLength 200 */
  website?: string;
}