# lib/api-zod/src/generated/types/sloperaWaitlistInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SloperaWaitlistInputProductId } from './sloperaWaitlistInputProductId';

export interface SloperaWaitlistInput {
  /** @maxLength 254 */
  email: string;
  productId: SloperaWaitlistInputProductId;
  requestId: string;
  /** @maxLength 200 */
  website?: string;
}